pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "notebooked"
version = "0.1.0"
description = "A tool to convert Jupyter notebooks into production-ready scripts and deploy them."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "boto3>=1.28.0",
    "sagemaker>=2.190.0",
    "mlflow>=2.8.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "pipreqs>=0.4.13",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
azure = [
    "azure-ai-ml>=1.0.0",
    "azure-identity>=1.14.0",
]
gcp = [
    "google-cloud-aiplatform>=1.36.0",
]

[project.scripts]
notebooked = "notebooked.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
